fixes visualize_shape_model_3d
