+++ menpowidgets/menpo3dwidgets/__init__.py
@@ -1,1 +1,1 @@
-from .base import visualize_shape_model_3d, visualize_morphable_model
+from .base import visualize_morphable_model
